requirements.txt
fastapi
uvicorn
groq
sentence-transformers
python-dotenv
pandas
scikit-learn
numpy
mangum
